Change play function to take double instead of integer

+++ src/main/scala/music2/util/Output.scala
@@ -1,7 +1,7 @@
 package music2.util
 
 object Output {
-  def play(x: Int) = {
+  def play(x: Double) = {
     ???
   }
 }
